requirements.txt
openai
python-dotenv==1.0.0
pymongo[srv]
streamlit

## Utils
ipykernel
